// src/enums/ZoneDefinition.ts
enum ZoneDefinition {
  Disabled = 48, // '0'
  BurglarEntryExit1 = 49, // '1'
  BurglarEntryExit2 = 50, // '2'
  BurglarPerimeterInstant = 51, // '3'
  BurglarInterior = 52, // '4'
  BurglarInteriorFollower = 53, // '5'
  BurglarInteriorNight = 54, // '6'
  BurglarInteriorNightDelay = 55, // '7'
  Burglar24Hour = 56, // '8'
  BurglarBoxTamper = 57, // '9'
  FireAlarm = 58, // ':'
  FireVerified = 59, // ';'
  FireSupervisory = 60, // '<'
  AuxAlarm1 = 61, // '='
  AuxAlarm2 = 62, // '>'
  Keyfob = 63, // '?'
  NonAlarm = 64, // '@'
  CarbonMonoxide = 65, // 'A'
  EmergencyAlarm = 66, // 'B'
  FreezeAlarm = 67, // 'C'
  GasAlarm = 68, // 'D'
  HeatAlarm = 69, // 'E'
  MedicalAlarm = 70, // 'F'
  PoliceAlarm = 71, // 'G'
  PoliceNoIndication = 72, // 'H'
  WaterAlarm = 73, // 'I'
  KeyMomentaryArmDisarm = 74, // 'J'
  KeyMomentaryArmAway = 75, // 'K'
  KeyMomentaryArmStay = 76, // 'L'
  KeyMomentaryDisarm = 77, // 'M'
  KeyOnOff = 78, // 'N'
  MuteAudibles = 79, // 'O'
  PowerSupervisory = 80, // 'P'
  Temperature = 81, // 'Q'
  AnalogZone = 82, // 'R'
  PhoneKey = 83, // 'S'
  IntercomKey = 84 // 'T'
}

export default ZoneDefinition;